refactor: change structure products
