+++ src/components/main/home/Catalog.tsx
@@ -1,214 +1,24 @@
 import React, { Component } from 'react'
-import { ProductCard } from './Productcard'
-
-import img1 from '../../../assets/catalog/1.jpg'
-import img2 from '../../../assets/catalog/2.jpg'
-import img3 from '../../../assets/catalog/3.jpg'
-import img4 from '../../../assets/catalog/4.jpg'
-import img5 from '../../../assets/catalog/5.jpg'
-import img6 from '../../../assets/catalog/6.jpg'
-import img7 from '../../../assets/catalog/7.jpg'
-import img8 from '../../../assets/catalog/8.jpg'
-import img9 from '../../../assets/catalog/9.jpg'
-import img10 from '../../../assets/catalog/10.jpg'
-import img11 from '../../../assets/catalog/11.jpg'
-import img12 from '../../../assets/catalog/12.jpg'
-import img13 from '../../../assets/catalog/13.jpg'
-import img14 from '../../../assets/catalog/14.jpg'
-import img15 from '../../../assets/catalog/15.jpg'
-import img16 from '../../../assets/catalog/16.jpg'
-import img17 from '../../../assets/catalog/17.jpg'
-import img18 from '../../../assets/catalog/18.jpg'
-import img19 from '../../../assets/catalog/19.jpg'
-import img20 from '../../../assets/catalog/20.jpg'
-import img21 from '../../../assets/catalog/21.jpg'
-
-const PRODUCTS = [
-  {
-    id: 1,
-    name: 'House 1',
-    price: 32000000,
-    imageUrl: img1,
-    description: '6 bedroom end of terrace house',
-    address: 'Wilton Mews, London'
-  },
-  {
-    id: 2,
-    name: 'House 2',
-    price: 'POA',
-    imageUrl: img2,
-    description: '8 bedroom for sale',
-    address: 'Queen Anne\'s Gate, London'
-  },
-  {
-    id: 3,
-    name: 'House 3',
-    price: 30000000,
-    imageUrl: img3,
-    description: '5 bedroom for sale',
-    address: 'Eaton Square, London'
-  },
-  {
-    id: 4,
-    name: 'House 4',
-    price: 30000000,
-    imageUrl: img4,
-    description: '5 bedroom end of terrace house',
-    address: 'Eaton Square, London'
-  },
-  {
-    id: 5,
-    name: 'House 5',
-    price: 'POA',
-    imageUrl: img5,
-    description: '9 bedroom for sale',
-    address: 'Eaton Square, London'
-  },
-  {
-    id: 6,
-    name: 'House 6',
-    price: 30000000,
-    imageUrl: img6,
-    description: '5 bedroom for sale',
-    address: 'Chester Square, London'
-  },
-    {
-    id: 7,
-    name: 'House 7',
-    price: 29500000,
-    imageUrl: img7,
-    description: '6 bedroom end of terrace house',
-    address: 'Wilton Mews, London'
-  },
-  {
-    id: 8,
-    name: 'House 8',
-    price: 25000000,
-    imageUrl: img8,
-    description: '8 bedroom for sale',
-    address: 'Queen Anne\'s Gate, London'
-  },
-  {
-    id: 9,
-    name: 'House 9',
-    price: 30000000,
-    imageUrl: img3,
-    description: '5 bedroom for sale',
-    address: 'Eaton Square, London'
-  },
-  {
-    id: 10,
-    name: 'House 10',
-    price: 33300000,
-    imageUrl: img10,
-    description: '10 bedroom end of terrace house',
-    address: 'Wilton Mews, London'
-  },
-  {
-    id: 11,
-    name: 'House 11',
-    price: 20500000,
-    imageUrl: img11,
-    description: '4 bedroom for sale',
-    address: 'Queen Anne\'s Gate, London'
-  },
-  {
-    id: 12,
-    name: 'House 12',
-    price: 22000000,
-    imageUrl: img12,
-    description: '5 bedroom for sale',
-    address: 'Eaton Square, London'
-  },
-  {
-    id: 13,
-    name: 'House 13',
-    price: 20500000,
-    imageUrl: img13,
-    description: '4 bedroom for sale',
-    address: 'Queen Anne\'s Gate, London'
-  },
-  {
-    id: 14,
-    name: 'House 14',
-    price: 194600000,
-    imageUrl: img14,
-    description: '5 bedroom for sale',
-    address: 'Eaton Square, London'
-  },
-  {
-    id: 15,
-    name: 'House 15',
-    price: 21400000,
-    imageUrl: img15,
-    description: '8 bedroom end of terrace house',
-    address: 'Wilton Mews, London'
-  },
-  {
-    id: 16,
-    name: 'House 16',
-    price: 30500000,
-    imageUrl: img16,
-    description: '7 bedroom for sale',
-    address: 'Queen Anne\'s Gate, London'
-  },
-  {
-    id: 17,
-    name: 'House 17',
-    price: 27700000,
-    imageUrl: img17,
-    description: '6 bedroom for sale',
-    address: 'Eaton Square, London'
-  },
-  {
-    id: 18,
-    name: 'House 18',
-    price: 'POA',
-    imageUrl: img18,
-    description: '4 bedroom for sale',
-    address: 'Queen Anne\'s Gate, London'
-  },
-  {
-    id: 19,
-    name: 'House 19',
-    price: 255000000,
-    imageUrl: img19,
-    description: '10 bedroom for sale',
-    address: 'Eaton Square, London'
-  },
-  {
-    id: 20,
-    name: 'House 20',
-    price: 'POA',
-    imageUrl: img20,
-    description: '4 bedroom for sale',
-    address: 'Queen Anne\'s Gate, London'
-  },
-  {
-    id: 21,
-    name: 'House 21',
-    price: 305000000,
-    imageUrl: img21,
-    description: '9 bedroom for sale',
-    address: 'Chester Square, London'
-  }
-]
+import { Product } from './Product'
+import { ProductCard } from './ProductCard'
+import { DataStore } from '../../../store/DataStore'
 
 interface CatalogProps {
-  products: Product[]
+  dataStore: DataStore
 }
 
-class Catalog extends Component<CatalogProps, CatalogState> {
+class Catalog extends Component<CatalogProps> {
   constructor(props: CatalogProps) {
     super(props)
   }
 
   render() {
-    const { products } = this.props
+    const { dataStore } = this.props
+    const products = new DataStore().getProducts()
 
     return (
       <div className="product-list">
-        {PRODUCTS.map(product => (
+        {products.map(product => (
           <ProductCard
             key={product.id}
             product={product}
