+++ src/components/main/home/Product.tsx
@@ -1,0 +1,10 @@
+interface Product {
+  id: number
+  name: string
+  price: number
+  imageUrl: string
+  description: string
+  address: string
+}
+
+export {Product}
